refactor: result 패키지 추가
