+++ src/main/java/blackjack/domain/user/Player.java
@@ -1,10 +1,6 @@
 package blackjack.domain.user;
 
-import blackjack.domain.Result;
-
 public class Player extends User {
-
-    public static final String CANNOT_DETERMINE_RESULT = "승패를 계산할 수 없습니다";
 
     public Player(String name) {
         super(name);
